tests/wpt/mozilla/tests/bluetooth/getPrimaryServices/disconnect-called-before-with-uuid.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.two_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => {
        gattServer.disconnect();
        return promise_rejects_dom(t, 'NetworkError', gattServer.getPrimaryServices(heart_rate.name));
    });
}, 'disconnect() called before getPrimaryServices. Reject with NetworkError.');
</script>
